services/content/scripts/pyproject.toml
[project]
name = "omnilingo-content-lesson-generator"
version = "0.1.0"
description = "Offline lesson generation tools for OmniLingo content-service"
requires-python = ">=3.11"
dependencies = [
  "asyncpg>=0.29,<1",
  "boto3>=1.34,<2",
  "google-genai>=1.0,<2",
  "httpx>=0.27,<1",
  "jsonschema>=4.22,<5",
  "pydantic>=2.7,<3",
  "pyyaml>=6,<7",
]

[tool.ruff]
line-length = 100
target-version = "py311"
